# Conflicts:
#	cellxgene_schema_cli/cellxgene_schema/migrate.py

+++ cellxgene_schema_cli/scripts/gene_processing.py
@@ -2,8 +2,10 @@ import csv
 import gzip
 import hashlib
 import os
+import statistics
 import sys
 import urllib.request
+from collections import defaultdict
 from typing import Dict
 
 import gtf_tools
@@ -14,11 +16,12 @@ import env
 
 
 class GeneProcessingResult:
-    def __init__(self, gene_id: str, gene_name: str, gene_version: str, gene_length: str):
+    def __init__(self, gene_id: str, gene_name: str, gene_version: str, gene_length: str, gene_type: str):
         self.gene_id = gene_id  # ex: ENSG00000141510, should be unique
         self.gene_name = gene_name  # ex: TP53, not necessarily unique
         self.gene_version = gene_version
         self.gene_length = gene_length
+        self.gene_type = gene_type
 
 
 class GeneProcessor:
@@ -72,14 +75,20 @@ class GeneProcessor:
 
                 line = line.rstrip().split("\t")  # type: ignore
 
-                # Desired features based on whether is gene or transcript
+                # Only process gene lines
                 if line[2] == "gene":
-                    features = ["gene_id", "gene_name", "gene_version"]
+                    # Attempt to fetch both gene_type and gene_biotype, since some ontologies use gene_type and others use gene_biotype
+                    features = ["gene_id", "gene_name", "gene_version", "gene_type", "gene_biotype"]
                 else:
                     continue
 
                 # Extract features (column 9 of GTF)
                 current_features = gtf_tools._get_features(line)  # type: ignore
+
+                # Set gene_name to gene_id if not present in GTF line
+                if "gene_name" not in current_features:
+                    current_features["gene_name"] = current_features["gene_id"]
+
                 # Filter genes suffixed with "PAR_Y"
                 if current_features["gene_id"].endswith("PAR_Y"):
                     continue
@@ -87,7 +96,7 @@ class GeneProcessor:
                 # get gene length
                 current_length = gene_lengths[current_features["gene_id"]]
 
-                # Select  features of interest, raise error if feature of interest not found
+                # Select features of interest, raise error if feature of interest not found
                 target_features = [""] * len(features)
                 for i in range(len(features)):
                     feature = features[i]
@@ -120,6 +129,8 @@ class GeneProcessor:
                     gene_name=target_features[1],
                     gene_version=target_features[2],
                     gene_length=str(current_length),
+                    # Prefer using gene_type, otherwise use gene_biotype
+                    gene_type=target_features[3] if target_features[3] != "" else target_features[4],
                 )
                 if gene_info_description in self.gene_ids_by_description:
                     self.gene_ids_by_description[gene_info_description].append(gene_id)
@@ -129,56 +140,45 @@ class GeneProcessor:
     def _get_gene_lengths_from_gtf(self, gtf_path: str) -> Dict[str, int]:
         """
         Parses a GTF file and calculates gene lengths, which are calculated as follows for each gene:
-        1. Get all different isoforms
-        2. Merge exons to create a set of non-overlapping "meta" exons
-        3. Sum the lengths of these "meta" exons
-
-        Code inspired from http://www.genemine.org/gtftools.php
+        1. Get lengths for all different isoforms
+        2. Get the median of the lengths of these isoforms
 
         :param str gtf_path: path to gzipped gtf file
 
         :rtype  Dict[str]
         :return A dictionary with keys being gene ids and values the corresponding length in base pairs
         """
-
+        gene_to_isoforms_map = defaultdict(set)
+        isoform_to_length_map = defaultdict(int)
         with gzip.open(gtf_path, "rb") as gtf:
-            # Dictionary of list of tuples that will store exon in bed format-like. Elements of the tuples will be the
-            # equivalent  fields from the bed format: chromosome, start, end, strand). Each list of tuples will correspond
-            # to one gene.
-
-            exons_in_bed = {}  # type: ignore
-
             for byte_line in gtf:
                 line = byte_line.decode("utf-8")
-
                 if line[0] == "#":
                     continue
 
-                line = line.rstrip().split("\t")  # type: ignore
+                # See https://www.gencodegenes.org/pages/data_format.html for GTF metadata schema
+                gene_metadata = line.rstrip().split("\t")  # type: ignore
 
-                if line[2] != "exon":
+                if gene_metadata[2] != "exon":
                     continue
 
-                # Convert line to bed-like format: (chromosome, start, end, strand)
-                exon_bed = (line[0], int(line[3]) - 1, int(line[4]), line[6])
-
-                current_features = gtf_tools._get_features(line)  # type: ignore
+                # Calculate exon length using genomic end location and genomic start location
+                exon_length = int(gene_metadata[4]) - int(gene_metadata[3]) + 1
+                current_features = gtf_tools._get_features(gene_metadata)  # type: ignore
                 gene_id = current_features["gene_id"]
-                if gene_id in exons_in_bed:
-                    exons_in_bed[gene_id].append(exon_bed)
-                else:
-                    exons_in_bed[gene_id] = [exon_bed]
+                transcript_id = current_features["transcript_id"]
 
-        # Merge exons from the same gene to create non-overlapping "meta" genes
-        # Then calculate gene length
+                gene_to_isoforms_map[gene_id].add(transcript_id)
+                isoform_to_length_map[transcript_id] += exon_length
+
         gene_lengths = {}
-        for gene in exons_in_bed:
-            meta_exons = gtf_tools.merge_bed_ranges(exons_in_bed[gene])  # type: ignore
-
-            # get length for this gene, i.e. sum of lengths of "meta" exons
-            gene_lengths[gene] = 0
-            for exon in meta_exons:
-                gene_lengths[gene] += exon[2] - exon[1]
+        for gene_id in gene_to_isoforms_map:
+            isoforms = gene_to_isoforms_map[gene_id]
+            isoform_lengths = []
+            for isoform in isoforms:
+                isoform_lengths.append(isoform_to_length_map[isoform])
+            # GTFTools established standard is to convert to int
+            gene_lengths[gene_id] = int(statistics.median(isoform_lengths))
 
         return gene_lengths
 
@@ -202,6 +202,7 @@ class GeneProcessor:
                     gene_name=ercc_id + " (spike-in control)",
                     gene_version=errc_version,
                     gene_length=errc_length,
+                    gene_type="synthetic",
                 )
                 if gene_info_description in self.gene_ids_by_description:
                     self.gene_ids_by_description[gene_info_description].append(ercc_id)
@@ -231,7 +232,7 @@ class GeneProcessor:
         # Write output for each file, and process file diffs
         for gene_info_key in gene_infos:
             gene_info_description = gene_infos[gene_info_key]["description"]
-            print("Writing output for ", gene_info_description)
+            print("Writing output for", gene_info_description)
             new_file = os.path.join(env.GENCODE_DIR, f"new_genes_{gene_info_description}.csv.gz")
             previous_ref_filepath = os.path.join(env.GENCODE_DIR, f"genes_{gene_info_description}.csv.gz")
             gene_ids = self.gene_ids_by_description[gene_info_description]
@@ -246,19 +247,26 @@ class GeneProcessor:
                                 gene_metadata.gene_name,
                                 gene_metadata.gene_version,
                                 gene_metadata.gene_length,
+                                gene_metadata.gene_type,
                             ]
                         )
                         + "\n"
                     )
                 self.write_gzip(output_to_print, new_file)
 
-                if self.digest(new_file) == self.digest(previous_ref_filepath):
-                    print("New gene reference is identical to previous gene reference", gene_info_description)
-                    os.remove(new_file)
-                else:
-                    print("generating gene reference diff for", gene_info_description)
-                    self.generate_gene_ref_diff(gene_info_description, new_file, previous_ref_filepath)
+                # Rename new_genes_{filename} to genes_{filename} if this is the first time we're introducing a new gene file type
+                if not os.path.exists(previous_ref_filepath):
                     os.replace(new_file, previous_ref_filepath)
+
+                # Process diff between new file and previous file, if there is a difference
+                else:
+                    if self.digest(new_file) == self.digest(previous_ref_filepath):
+                        print("New gene reference is identical to previous gene reference", gene_info_description)
+                        os.remove(new_file)
+                    else:
+                        print("generating gene reference diff for", gene_info_description)
+                        self.generate_gene_ref_diff(gene_info_description, new_file, previous_ref_filepath)
+                        os.replace(new_file, previous_ref_filepath)
             except Exception as e:
                 print("Writing to new file failed. Using previous version.", gene_info_description)
                 raise e
